+++ app/src/modules/settings/routes/translation-strings/translation-strings-drawer.vue
